+++ lib/Container.d.ts
@@ -200,6 +200,7 @@ export default class Container extends EventEmitter {
      *  - add to the devices map
      *  - create the structure entry
      *  - run `preProcess()`
+     *  - attach auto-render of `shares` (`attachSharesRender()`)
      *  - register actions
      *  - register metrics (emit `device.register.metric`)
      *  - create input & output ports